.claude/worktrees/heuristic-hamilton/requirements.txt
matplotlib>=3.5
numpy>=1.21
PyYAML>=6.0
